moved maybeAlertForPipelineRun into its own file
